+++ src/main/java/me/marin/lockout/Constants.java
@@ -24,6 +24,8 @@ public class Constants {
     public static final Identifier SET_BOARD_TYPE_PACKET = Identifier.of(NAMESPACE, "set_board_type");
     public static final Identifier UPLOAD_BOARD_TYPE_PACKET = Identifier.of(NAMESPACE, "upload_board_type");
     public static final Identifier ANNOUNCE_GOAL_FOCUS_PACKET = Identifier.of(NAMESPACE, "announce_goal_focus");
+    public static final Identifier REQUEST_GOAL_DETAILS_PACKET = Identifier.of(NAMESPACE, "request_goal_details");
+    public static final Identifier GOAL_DETAILS_PACKET = Identifier.of(NAMESPACE, "goal_details");
     
     public static final Identifier START_PICK_BAN_SESSION_PACKET = Identifier.of(NAMESPACE, "start_pick_ban_session");
     public static final Identifier UPDATE_PICK_BAN_SESSION_PACKET = Identifier.of(NAMESPACE, "update_pick_ban_session");